[#S12P11B109-번호] 무작위 초밥 제공 api

See merge request s12-webmobile2-sub1/S12P11B109!29

+++ backend/sushi/src/main/java/com/ssafy/sushi/domain/user/UserService.java
@@ -18,7 +18,7 @@ public class UserService {
     private final UserRepository userRepository;
 
 
-    public UserInfoResponse getUserInfo(Long userId) {
+    public UserInfoResponse getUserInfo(Integer userId) {
         User user = userRepository.findById(userId).orElseThrow(() ->
                 new CustomException(ErrorCode.USER_NOT_FOUND));
