+++ backend/sushi/src/main/java/com/ssafy/sushi/global/common/util/AuthenticationUtil.java
@@ -5,14 +5,14 @@ import com.ssafy.sushi.global.error.exception.CustomException;
 import com.ssafy.sushi.global.security.UserPrincipal;
 
 public class AuthenticationUtil {
-    public static Long getCurrentUserId(UserPrincipal userPrincipal) {
+    public static Integer getCurrentUserId(UserPrincipal userPrincipal) {
         if (userPrincipal == null) {
             throw new CustomException(ErrorCode.USER_NOT_FOUND);
         }
         return userPrincipal.getId();
     }
 
-    public static Long getCurrentUserIdCanNull(UserPrincipal userPrincipal) {
+    public static Integer getCurrentUserIdCanNull(UserPrincipal userPrincipal) {
         return userPrincipal == null ? null : userPrincipal.getId();
     }
 }
